requirements.txt
Flask==2.2.2
line-bot-sdk==2.3.0
beautifulsoup4==4.9.3
soupsieve==2.2.1
urllib3==1.26.4